[GUI General] Implemented an example race UI in java/com/group17/hifiprototype/scenes/f1_scene Added instructions for adding new scenes to README.md
[GUI General]
Added the Formula 1 and the Main page layouts.
Some further layout is still needed alongside data and styling

[GUI General]
Improved Formula 1 GUI to include the flag and the buttons that will be used.

[GUI General]
Added Back button and overlay open/close to F1 GUI.
Work needs to be done to implement the scrollable parts.

[GUI General]
Added most of the scrollpanes, but resizing kills the program.

[GUI General]
Added ScrollPane utility functions for the overlay VBox.
 Resizing still kills the program.

[GUI General]
Just a save in case I mess something up

[GUI General]
Added height/width resizing to the bottom scrollbar.
Capped how much it can be dragged up

[GUI General]
Fixed resizing problem.
The VBoxControls code is a mess, but it works.

[GUI General]
Tidied up VBoxControls

[GUI General]
Refactored code and added documentation for the functions in the code. (No readmes, sorry.)

[Docs]
Added a README file for adding new scenes